+++ specs/binding-kafka.spec/src/test/java/io/aklivity/zilla/specs/binding/kafka/streams/network/KafkaDeleteTopicsApiVersionsIT.java
@@ -27,15 +27,11 @@ import org.junit.rules.Timeout;
 import io.aklivity.k3po.runtime.junit.annotation.Specification;
 import io.aklivity.k3po.runtime.junit.rules.K3poRule;
 
-public class CreateTopicsApiVersionsIT
+public class KafkaDeleteTopicsApiVersionsIT
 {
     private final K3poRule k3po = new K3poRule()
         .addScriptRoot("net",
-            "io/aklivity/zilla/specs/binding/kafka/streams/network/create.topics.v3.api.versions.v0")
-        .addScriptRoot("netUnsupported",
-            "io/aklivity/zilla/specs/binding/kafka/streams/network/create.topics.v3.api.versions.v0.unsupported")
-        .addScriptRoot("netReused",
-            "io/aklivity/zilla/specs/binding/kafka/streams/network/create.topics.v3.api.versions.v0.reused");
+            "io/aklivity/zilla/specs/binding/kafka/streams/network/delete.topics.v3.api.versions.v0");
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(5, SECONDS));
 
@@ -44,27 +40,9 @@ public class CreateTopicsApiVersionsIT
 
     @Test
     @Specification({
-        "${net}/create.topics/client",
-        "${net}/create.topics/server"})
-    public void shouldCreateTopicsV3ApiVersionsV0() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${netUnsupported}/create.topics/client",
-        "${netUnsupported}/create.topics/server"})
-    public void shouldCreateTopicsV3UnsupportedVersion() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${netReused}/create.topics.then.delete.topics/client",
-        "${netReused}/create.topics.then.delete.topics/server"})
-    public void shouldCreateTopicsV3ReusedConnection() throws Exception
+        "${net}/delete.topics/client",
+        "${net}/delete.topics/server"})
+    public void shouldDeleteTopicsV3ApiVersionsV0() throws Exception
     {
         k3po.finish();
     }
